requirements.txt
simplejson
lxml
codecov
coverage
flask
pre - commit
pytest
PyJWT
requests
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
Werkzeug==2.0.0
